+++ tutorials/sphinx-tutorials/torchrl_demo.py
@@ -61,7 +61,7 @@ This demo was presented at ICML 2022 on the industry demo day.
 #       │   │   └── "samplers.py"
 #       │   │   └── "storages.py"
 #       │   │   └── "writers.py"
-#       │   ├── "rlhf"
+#       │   ├── "llm"
 #       │   │   └── "dataset.py"
 #       │   │   └── "prompt.py"
 #       │   │   └── "reward.py"
@@ -91,7 +91,7 @@ This demo was presented at ICML 2022 on the industry demo day.
 #       │   │   └── "functional.py"
 #       │   │   └── "gym_transforms.py"
 #       │   │   └── "r3m.py"
-#       │   │   └── "rlhf.py"
+#       │   │   └── "llm.py"
 #       │   │   └── "vc1.py"
 #       │   │   └── "vip.py"
 #       │   └── "vec_envs.py"
@@ -106,7 +106,7 @@ This demo was presented at ICML 2022 on the industry demo day.
 #       │   │   └── "model_based.py"
 #       │   │   └── "models.py"
 #       │   │   └── "multiagent.py"
-#       │   │   └── "rlhf.py"
+#       │   │   └── "llm.py"
 #       │   ├── "planners"
 #       │   │   └── "cem.py"
 #       │   │   └── "common.py"
